Move SidebarScrollViewResolverView to its own file
Autoreview: the extraction from ContentView.swift should follow the
one-major-type-per-file architecture rule rather than sharing
SidebarScrollViewConfigurator.swift with the configurator enum. No
behavior change; wires the new file into the app target.
